refactor(onboard): converge on shared isGatewayHttpReady (#3312) + fix md lint
Address PR review findings:

1. Merge origin/main to pick up PR #3312 (isGatewayHttpReady from
   src/lib/onboard/gateway-http-readiness.ts). Replace our ad-hoc
   verifyDockerDriverGatewayListening (TCP probe) with a call to the
   shared isGatewayHttpReady helper — HTTP is strictly stronger than
   a raw TCP connect and is already the de-facto standard used by
   every other gateway-reuse decision site in onboard.ts after #3312.
   Docker-driver and K3s paths now converge on the same probe.

2. Drop the parallel test/gateway-tcp-probe.test.ts (9 tests for a
   helper that no longer exists). The shared helper's behavior is
   already covered by test/gateway-http-reuse-wait.test.ts (21 tests).
   Replace with test/gateway-health-honest-integration.test.ts (6
   source-shape tests) guarding the #3111 integration pattern.

3. Fix 7 markdownlint errors in ACCEPTANCE.md (MD040, MD022, MD031) —
   the 'checks' job flagged these; now passes locally.

4. Update ACCEPTANCE.md to reflect the simpler design: Phase 1 is
   'reuse existing helper' rather than 'add new TCP probe helper',
   and the refactoring alignment table records #3312 as 'adopted'
   rather than 'to be coordinated'.

Behavior of the fix is unchanged:
  - child.once('exit') listener tracks zombies (detached children
    that process.kill(pid, 0) would falsely report as alive)
  - poll loop guard is childExited || !isPidAlive(childPid)
  - healthy log gated on isGatewayHttpReady() after isGatewayHealthy()
  - user-visible exit code/signal surfaced on failure

Related: #3111
Coverage guard: #3362 — gateway-health-honest-e2e
