+++ src/lib/onboard.ts
@@ -3953,47 +3953,6 @@ function isPidAlive(pid: number): boolean {
   }
 }
 
-/**
- * Probe a TCP endpoint to verify something is actually listening and
- * accepting connections. Used to gate the "gateway is healthy" log in
- * startDockerDriverGateway against the class of bug reported in #3111 where
- * the openshell-gateway binary crashed on startup but metadata-based health
- * checks (isGatewayHealthy) still returned true.
- *
- * Resolves true on successful TCP connect. Resolves false on connection
- * refused, unreachable host, or timeout. Never rejects.
- *
- * TODO(#2562): adopt the unified timeout abstraction once that epic lands.
- */
-async function verifyDockerDriverGatewayListening(
-  port: number,
-  timeoutMs = 500,
-  host = "127.0.0.1",
-): Promise<boolean> {
-  // Lazy-require so tests that don't exercise the gateway path don't pay
-  // for the net module load, and so the import doesn't fan out across the
-  // already-bloated top-of-file require block.
-  // biome-ignore lint/suspicious/noExplicitAny: node:net types are not in scope in this file
-  const net: any = require("node:net");
-  return new Promise<boolean>((resolve) => {
-    let settled = false;
-    const settle = (result: boolean) => {
-      if (settled) return;
-      settled = true;
-      try {
-        socket.destroy();
-      } catch {
-        /* best effort */
-      }
-      resolve(result);
-    };
-    const socket = net.createConnection({ host, port });
-    socket.setTimeout(Math.max(50, timeoutMs));
-    socket.once("connect", () => settle(true));
-    socket.once("timeout", () => settle(false));
-    socket.once("error", () => settle(false));
-  });
-}
 
 function getDockerDriverGatewayPid(): number | null {
   try {
@@ -5482,11 +5441,10 @@ async function startDockerDriverGateway({
 
   const pollCount = envInt("NEMOCLAW_HEALTH_POLL_COUNT", 30);
   const pollInterval = envInt("NEMOCLAW_HEALTH_POLL_INTERVAL", 2);
-  const tcpProbeTimeoutMs = envInt("NEMOCLAW_GATEWAY_TCP_PROBE_TIMEOUT_MS", 500);
   for (let i = 0; i < pollCount; i += 1) {
-    // #3111: combine the exit-event check with isPidAlive. exit-event catches
-    // zombies that isPidAlive misses; isPidAlive catches children we lost
-    // track of (e.g., if the exit listener was removed).
+    // #3111: combine the exit-event check with isPidAlive. The exit event
+    // catches zombies that isPidAlive misses; isPidAlive catches children
+    // we somehow lost track of (defense in depth).
     if (childExited || !isPidAlive(childPid)) {
       break;
     }
@@ -5501,17 +5459,19 @@ async function startDockerDriverGateway({
     const currentInfo = runCaptureOpenshell(["gateway", "info"], { ignoreError: true });
     if (isGatewayHealthy(status, namedInfo, currentInfo)) {
       // #3111: isGatewayHealthy is a pure string match on openshell CLI
-      // output. "Server Status" is a header emitted unconditionally, even
-      // when the gateway endpoint is not reachable (Connection refused).
-      // Gate the "healthy" log on a real TCP probe so we don't claim
-      // healthy for a dead or zombied gateway. On probe failure, keep
-      // polling — the binary may still be binding its listener. The
-      // childExited / isPidAlive check at the top of the loop will
-      // terminate us if the process actually died.
+      // output. isGatewayConnected matches on "Server Status" which is a
+      // header the CLI prints unconditionally, so the metadata check can
+      // return true even when the gateway endpoint is unreachable
+      // (Connection refused). Gate the "healthy" log on a real HTTP probe
+      // via isGatewayHttpReady() — the same helper used at every other
+      // gateway-reuse decision site after #3312. On probe failure, keep
+      // polling: the binary may still be binding its listener. The
+      // childExited / isPidAlive check at the top of the loop terminates
+      // us if the process actually died.
       //
       // TODO(#3213): surface this as a structured advisory rather than
       // plain log output once the unified registry lands.
-      if (await verifyDockerDriverGatewayListening(GATEWAY_PORT, tcpProbeTimeoutMs)) {
+      if (await isGatewayHttpReady()) {
         console.log("  ✓ Docker-driver gateway is healthy");
         return;
       }
@@ -12093,7 +12053,6 @@ module.exports = {
   getResumeConfigConflicts,
   isGatewayHealthy,
   hasStaleGateway,
-  verifyDockerDriverGatewayListening,
   getRequestedSandboxNameHint,
   getResumeSandboxConflict,
   getSandboxReuseState,
